+++ libraries/splunk_app_provider.rb
@@ -83,7 +83,8 @@ class Chef
         if new_resource.templates
           new_resource.templates.each do |t|
             template "#{app_dir}/local/#{t}" do
-              source "#{new_resource.app_name}/#{t}.erb"
+              source "#{new_resource.app_name}/#{t}"
+              cookbook new_resource.cookbook
               owner splunk_runas_user
               group splunk_runas_user
               mode '644'